+++ runtime/src/ujima/desktop/app/catalog.clj
@@ -31,6 +31,7 @@
 (defn init! [catalog] 
   (reset! catalog* catalog))
 
+
 (defn current [] @catalog*)